third_party/blink/web_tests/wpt_internal/geolocation-api/error-service-connection-error.https.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/testharness-adapter.js"></script>
</head>
<body>
<script type="module">
import {GeolocationMock} from './resources/geolocation-mock.js';

description("Tests Geolocation when the geolocation service connection fails.");

let error;

const mock = new GeolocationMock();
mock.setGeolocationPermission(true);
mock.rejectGeolocationServiceConnections();

navigator.geolocation.getCurrentPosition(function(p) {
  testFailed('Success callback invoked unexpectedly');
  finishJSTest();
}, function(e) {
  error = e;
  assert_equals(error.code, error.PERMISSION_DENIED);
  assert_equals(error.message, 'User denied Geolocation');
  finishJSTest();
});
</script>
</body>
</html>
